refactor(complaints): extrai lógica de fetch para hook customizado
- Move lógica de fetch do Context para hook useComplaintsFetch
- Reduz ComplaintsContext de 98 para 54 linhas (44% menor)
- Cria hook reutilizável para fetch de denúncias
- Extrai lógica repetida de AbortController para função auxiliar
- Melhora separação de responsabilidades e testabilidade
- Mantém mesma funcionalidade com código mais organizado
